requirements.txt
discord.py>=2.3
APScheduler>=3.10
python-dotenv>=1.0
